Move backend object files to be with their backends
