+++ claripy/ast/bv.py
@@ -312,7 +312,7 @@ def SI(
 ):
     name = "unnamed" if name is None else name
     if to_conv is not None:
-        si = claripy.vsa.CreateStridedInterval(
+        si = claripy.backends.backend_vsa.CreateStridedInterval(
             name=name, bits=bits, lower_bound=lower_bound, upper_bound=upper_bound, stride=stride, to_conv=to_conv
         )
         return BVS(
@@ -352,7 +352,7 @@ def ValueSet(bits, region=None, region_base_addr=None, value=None, name=None, va
     if isinstance(v, numbers.Number):
         min_v, max_v = v, v
         stride = 0
-    elif isinstance(v, claripy.vsa.StridedInterval):
+    elif isinstance(v, claripy.backends.backend_vsa.StridedInterval):
         min_v, max_v = v.lower_bound, v.upper_bound
         stride = v.stride
     elif isinstance(v, claripy.ast.Base):
